requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1

# AI and LLM
openai==1.10.0
anthropic==0.18.0
google-generativeai==0.3.2

# Vector Database and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
